Output calculated checksum to `stdout`.
