+++ liquibase-standard/src/main/java/liquibase/command/core/ExecuteSqlCommandStep.java
@@ -1,6 +1,5 @@
 package liquibase.command.core;
 
-import liquibase.GlobalConfiguration;
 import liquibase.Scope;
 import liquibase.changelog.ChangeLogParameters;
 import liquibase.changeset.ChangeSetService;
@@ -20,8 +19,6 @@ import liquibase.util.StreamUtil;
 import liquibase.util.StringUtil;
 
 import java.io.IOException;
-import java.io.OutputStreamWriter;
-import java.io.Writer;
 import java.util.Arrays;
 import java.util.LinkedHashSet;
 import java.util.List;
@@ -98,13 +95,6 @@ public class ExecuteSqlCommandStep extends AbstractCommandStep {
             delimiter = service.getEndDelimiter(null);
         }
         return delimiter;
-    }
-
-    protected void handleOutput(CommandResultsBuilder resultsBuilder, String output) throws IOException {
-        String charsetName = GlobalConfiguration.OUTPUT_FILE_ENCODING.getCurrentValue();
-        Writer outputWriter = new OutputStreamWriter(resultsBuilder.getOutputStream(), charsetName);
-        outputWriter.write(output);
-        outputWriter.flush();
     }
 
     protected String getSqlScript(String sql, String sqlFile) throws IOException, LiquibaseException {
